HardWay2D(Developed with Unity)
Developed with Unity 2d. The game content includes an ambulance, various car obstacles, a health bar and time. The aim of the game is to complete the time without depleting the ambulance's health value in 60 seconds. There are difficulty level and sound settings. The spawn speed of car obstacles varies according to the difficulty level.